Fix iOS build errors with platform conditionals
- Wrap macOS-only views (MenuBarView, AboutView) in #if os(macOS)
- Wrap TokenUsageService in #if os(macOS)
- Add separate iOS initializer for UsageViewModel (no tokenService)
- Move macOS-only Constants (credentialsFileURL, claudeProjectsDirectories) to #if os(macOS)
- Update iOS app entry to use simplified initializer
- Remove unused Xcode template files
