+++ ClaudeMeter-iOS/ClaudeMeter_iOSApp.swift
@@ -2,31 +2,26 @@
 //  ClaudeMeter_iOSApp.swift
 //  ClaudeMeter-iOS
 //
-//  Created by Ru Chern Chong on 1/1/26.
-//
 
 import SwiftUI
-import SwiftData
 
 @main
 struct ClaudeMeter_iOSApp: App {
-    var sharedModelContainer: ModelContainer = {
-        let schema = Schema([
-            Item.self,
-        ])
-        let modelConfiguration = ModelConfiguration(schema: schema, isStoredInMemoryOnly: false)
+    @State private var viewModel: UsageViewModel
 
-        do {
-            return try ModelContainer(for: schema, configurations: [modelConfiguration])
-        } catch {
-            fatalError("Could not create ModelContainer: \(error)")
-        }
-    }()
+    init() {
+        let credentialService = iOSCredentialService()
+        _viewModel = State(initialValue: UsageViewModel(
+            credentialProvider: credentialService
+        ))
+    }
 
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            NavigationStack {
+                DashboardView()
+                    .environment(viewModel)
+            }
         }
-        .modelContainer(sharedModelContainer)
     }
 }
